Replaced all multiplications in the algorithm
Current max frequency: 45.5MHz.

-The multiplication involving MU in the previous commit has been
replaced by the function negate_multiply. The function is now called
mu_multiply in this commit.

-The multiplication between delta and X has been replaced by a sll
operation. The process which updates delta's value has been changed to
return a new integer signal called delta_pos, which is the bit position
of the 1's location in delta.